feature: oc_clean_istags deletes tags pointing at unreachable commits

+++ lib/oc_clean_istags.sh
@@ -18,14 +18,17 @@ the development process:
 Goes through all the imagestream tags that exists on the provided imagestream,
 and deletes the following tags:
  - tags that point to a git object (commit or branch) which
-   do not exist in the current git repository
+   do not exist in the current git repository, or is unreachable
  - tags that point to a git commit which is available in the "develop" branch,
    but is not a merge commit
+
+The remaining commits (merge commits on develop or master) are then listed,
+and the 20 most recent tags are kept, older tags are deleted.
 
 If a git commit has a git tag pointing to it,
 the corresponding imagestream tag will never be deleted
 
-Usage: $0 OC_COMMAND PROJECT IMAGE_STREAM_NAME GIT_EXECUTABLE...
+Usage: $0 OC_COMMAND PROJECT IMAGE_STREAM_NAME GIT_EXECUTABLE ...
 EOF
     exit 1
 }
@@ -57,15 +60,22 @@ do
         # Delete tags pointing to objects that do not exist anymore
         # both commits and branches return the type "commit"
         $oc -n "$oc_project" delete istag/"$image_stream":"$istag"
+        echo "$istag doesn't exist in the repo"
+    elif [[ "$($git cat-file -t "$istag" 2>/dev/null)" == "commit" && -z $($git branch --contains "$istag") ]]; then
+        # Delete tags pointing to commits that are unreacheable
+        $oc -n "$oc_project" delete istag/"$image_stream":"$istag"
+        echo "$istag was unreachable"
     elif $git branch --contains "$istag" | grep -q -E '(develop|master)'; then
         # The tag is on develop or master
-        if [ "$($git log --pretty=%P -n 1 "$istag" | wc -w)" -eq 1 ]; then
+        if [ "$($git log --pretty=%P -n 1 "$istag" | wc -w)" == "1" ]; then
             # the tag is now in develop or master but is not a merge commit
             if [ -z "$($git tag --points-at "$istag" 2>/dev/null)" ]; then # never delete imagestream tag if there is a git tag pointing at it
                 $oc -n "$oc_project" delete istag/"$image_stream":"$istag"
+                echo "$istag is not a merge commit"
+            else
+                echo "Keeping $image_stream:$istag because a Git tag points at it"
             fi;
-        fi
-        if [[ ! "${normalized_branches[*]}" =~ (^| )"${istag}"( |$) ]]; then
+        elif [[ ! "${normalized_branches[*]}" =~ (^| )"${istag}"( |$) ]]; then
             # The tag is not a branch. Candidate for deletion based on age
             echo "${istag}" >> "$kept_istags"
         fi
@@ -79,7 +89,11 @@ git rev-list master | while read -r sha1; do
         if [ $counter -le 20 ]; then
             echo "keeping $sha1"
         else
-            $oc -n "$oc_project" delete istag/"$image_stream":"$sha1"
+            if [ -z "$($git tag --points-at "$istag" 2>/dev/null)" ]; then
+                $oc -n "$oc_project" delete istag/"$image_stream":"$sha1"
+            else
+                echo "Keeping $image_stream:$istag because a Git tag points at it"
+            fi;
         fi
     fi
 done
